requirements.txt
alembic==1.10.2
anyio==3.6.2
click==8.1.3
fastapi==0.94.1
FastAPI-SQLAlchemy==0.2.1
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
psycopg2==2.9.5
pydantic==1.10.6
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
yapf==0.32.0
